Move (Default)ResolutionProvider from reporter to model
These classes only depend on the model, and the move is consistent with
other classes like PackageConfigurationProvider in the destination.
Finally, this allows to remove the dependency of helper-cli on reporter.
