+++ helper-cli/build.gradle.kts
@@ -57,7 +57,6 @@ repositories {
 dependencies {
     implementation(project(":analyzer"))
     implementation(project(":downloader"))
-    implementation(project(":reporter"))
     implementation(project(":scanner"))
     implementation(project(":utils"))
