+++ reporter/src/funTest/kotlin/reporters/EvaluatedModelReporterTest.kt
@@ -20,7 +20,7 @@
 package org.ossreviewtoolkit.reporter.reporters
 
 import org.ossreviewtoolkit.model.OrtResult
-import org.ossreviewtoolkit.reporter.DefaultResolutionProvider
+import org.ossreviewtoolkit.model.utils.DefaultResolutionProvider
 import org.ossreviewtoolkit.reporter.ReporterInput
 import org.ossreviewtoolkit.utils.normalizeLineBreaks
 import org.ossreviewtoolkit.utils.test.readOrtResult
